# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Tienda",
        "description": "Esta es la documentación de la API que permite gestionar categorías, productos y órdenes en una tienda.",
        "contact": {
            "email": "jmcabo93@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/order-items": {
            "get": {
                "tags": [
                    "OrderItem"
                ],
                "summary": "Obtener los items de una orden",
                "description": "Obtiene todos los items de una orden con los productos relacionados",
                "operationId": "00ef4347ae4117974e08654982418127",
                "responses": {
                    "200": {
                        "description": "Lista de items de la orden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderItem"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No hay items en esta orden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No hay items en esta orden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error al obtener los items de la orden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "OrderItem"
                ],
                "summary": "Almacenar un nuevo item en una orden",
                "description": "Crea un nuevo item en la orden, verificando el stock disponible y creando la relación con el producto",
                "operationId": "ebd626997709bc1943529afd1010f1b2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderItem"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item de la orden creado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderItem"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Stock insuficiente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Stock insuficiente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al crear el item de la orden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error al crear el item de la orden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/order-items/{id}": {
            "put": {
                "tags": [
                    "OrderItem"
                ],
                "summary": "Actualizar un item de la orden",
                "description": "Actualiza los detalles de un item en la orden, verificando el stock disponible",
                "operationId": "75cbe80759b818ea008f79628f18c87e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del item de la orden",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderItem"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item de la orden actualizado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderItem"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Stock insuficiente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Stock insuficiente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al actualizar el item de la orden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error al actualizar el item de la orden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "OrderItem"
                ],
                "summary": "Eliminar un item de la orden",
                "description": "Elimina un item de la orden y restaura el stock del producto",
                "operationId": "1278a6534795039ad2102b83915b6ead",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del item de la orden",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Item de la orden eliminado con éxito"
                    },
                    "500": {
                        "description": "Error al eliminar el item de la orden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error al eliminar el item de la orden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrderItem": {
                "required": [
                    "order_id",
                    "product_id",
                    "quantity",
                    "price"
                ],
                "properties": {
                    "id": {
                        "description": "ID del item de la orden",
                        "type": "integer",
                        "example": 1
                    },
                    "order_id": {
                        "description": "ID de la orden",
                        "type": "integer",
                        "example": 101
                    },
                    "product_id": {
                        "description": "ID del producto",
                        "type": "integer",
                        "example": 202
                    },
                    "quantity": {
                        "description": "Cantidad de productos",
                        "type": "integer",
                        "example": 2
                    },
                    "price": {
                        "description": "Preciototal",
                        "type": "number",
                        "format": "float",
                        "example": 50
                    },
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    }
                },
                "type": "object"
            },
            "Product": {
                "required": [
                    "name",
                    "description",
                    "price",
                    "stock",
                    "category_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID del producto",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nombre del producto",
                        "type": "string",
                        "example": "Smartphone"
                    },
                    "description": {
                        "description": "Descripción del producto",
                        "type": "string",
                        "example": "Electronics"
                    },
                    "price": {
                        "description": "Precio del producto",
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "stock": {
                        "description": "Cantidad de stock disponible",
                        "type": "integer",
                        "example": 200
                    },
                    "category_id": {
                        "description": "ID de la categoría a la que pertenece el producto",
                        "type": "integer",
                        "example": 1
                    },
                    "image": {
                        "description": "Imagen del producto",
                        "type": "string",
                        "example": "",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Usa un token JWT en formato Bearer para autenticarte",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}